allow more detailed view of jobs
Added more trace logs about the status of the sge
execution. This now includes the output of the
'qsub' command (under the trace debug level).
We now use a new function, sge.system.tee, that
prints the output from a command and also returns
it. Here's a word of explanation for why we made
this change (instead of using the 'system'
function). In our setting, we set the sge.qsub
option to '/usr/share/doc/expect-dev/examples/unbuffer qsub'.
This "unbuffer" script, available in Ubuntu/debian
linux in the expect-dev package allows you to run
qsub in an unbuffered way, so that we can monitor
the completion and return status of the different
jobs as they complete. Of course, everything works
fine if you keep sge.qsub as 'qsub'.

(this patch was originally submitted by
Gal GoldShtein and Alon Wasserman)